requirements.txt
absl-py==2.1.0
attrs==24.2.0
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.55.1
fsspec==2024.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.3
mediapipe==0.10.18
ml_dtypes==0.5.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
pycparser==2.22
pygame==2.6.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.5.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
typing_extensions==4.9.0
tzdata==2024.2
